requirements.txt
# Core trading dependencies
ccxt==4.5.2
pandas==2.3.2
numpy==1.26.4
pandas-ta==0.3.14b
vectorbt==0.28.1
python-dotenv==1.1.1
matplotlib==3.10.6
typer==0.17.3
pydantic==2.11.7
loguru==0.7.3
pyyaml==6.0.2

# Real-time dashboard dependencies
fastapi==0.115.6
uvicorn==0.34.0
websockets==14.1
aiohttp==3.11.17

# Development tools (optional in production)
black==25.1.0
ruff==0.12.11
pytest==8.4.1
jupyter==1.1.1
ipykernel==6.30.1